+++ demos/chinook/src/main/java/is/codion/demos/chinook/ui/AlbumTablePanel.java
@@ -71,7 +71,7 @@ public final class AlbumTablePanel extends EntityTablePanel {
 
 	private void viewSelectedCover() {
 		tableModel().selection().item().optional()
-						.filter(album -> album.isNotNull(Album.COVER))
+						.filter(album -> !album.isNull(Album.COVER))
 						.ifPresent(album -> displayCover(album.get(Album.TITLE), album.get(Album.COVER)));
 	}